requirements.txt
pyyaml
ipython